fix logical errors; refactor

+++ src/main/java/ca/jrvs/apps/twitter/TwitterCLIRunner.java
@@ -9,13 +9,17 @@ import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.util.InputMismatchException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Scanner;
 
 @Component
 public class TwitterCLIRunner {
-    private static final String COMMA = ",";
     private static final String COLON = ":";
+    private static final int POST_MIN_ARGS = 3;
+    private static final int SHOW_MIN_ARGS = 3;
+    private static final int DEL_MIN_ARGS = 2;
     private TwitterService service;
     private Scanner coordinateParser;
 
@@ -25,15 +29,12 @@ public class TwitterCLIRunner {
     }
 
     public void run(String[] args) {
-
         if (args.length < 1) {
             System.err.println(TwitterUtil.SERVICE_USAGE_TEMP);
             return;
         }
-
         TwitterServiceImp service = new TwitterServiceImp(new TwitterRestDao());
         String cmd = args[0];
-
         switch (cmd) {
             case "post":
                 postTweet(args);
@@ -52,13 +53,12 @@ public class TwitterCLIRunner {
 
     private void showTweet(String[] args) {
         String idStr = args[1];
-        if (args.length < 3) {
-            System.err.println(TwitterUtil.SERVICE_APP_USAGE);
-            return;
+        if (args.length < SHOW_MIN_ARGS) {
+            System.err.println(TwitterUtil.SERVICE_SHOW_USAGE); return;
         }
-        int fields_len = args.length - TwitterUtil.CMD_ARG_INDEX;
+        int fields_len = args.length - TwitterUtil.CMD_SHOW_ARG_INDEX;
         String[] fields = new String[fields_len];
-        System.arraycopy(args, TwitterUtil.CMD_ARG_INDEX, fields, 0, fields_len);
+        System.arraycopy(args, TwitterUtil.CMD_SHOW_ARG_INDEX, fields, 0, fields_len);
         try {
             service.showTweet(idStr, fields);
         } catch (IOException e) {
@@ -69,44 +69,52 @@ public class TwitterCLIRunner {
     }
 
     private void deleteTweet(String[] args) {
+        if (args.length < DEL_MIN_ARGS) {
+            System.err.println(TwitterUtil.SERVICE_DEL_USAGE); return;
+        }
         List<Tweet> tweets = null;
         String[] ids = new String[args.length - 1];
-        System.arraycopy(args, 2, ids, 0, args.length);
+        System.arraycopy(args, TwitterUtil.CMD_DEFAULT_ARG_INDEX, ids, 0, ids.length);
         try {
             tweets = service.deleteTweets(ids);
+            System.out.println("Successfully deleted: ");
+            tweets.forEach(t -> System.out.println(t.getIdStr()));
         } catch (IOException e) {
             e.printStackTrace();
         } catch (URISyntaxException e) {
             e.printStackTrace();
+        } catch (NoSuchElementException e) {
+            e.printStackTrace();
         }
-        if (tweets == null) {
-            System.err.println("Delete: " + TwitterUtil.INVALID_EX_MSG);
-        }
-        System.out.println("Successfully deleted: ");
-        tweets.forEach(t -> System.out.println(t.getIdStr()));
     }
 
     private Tweet postTweet(String[] args) {
+        if (args.length < POST_MIN_ARGS) {
+            System.err.println(TwitterUtil.SERVICE_POST_USAGE); return null;
+        }
         String text = args[1];
-        coordinateParser = getParser(args);
-        Double longitude = coordinateParser.nextDouble();
-        Double latitude = coordinateParser.nextDouble();
+        coordinateParser = getCoordinateParser(args);
         Tweet tweet = null;
         try {
+            Double longitude = coordinateParser.nextDouble();
+            Double latitude = coordinateParser.nextDouble();
             tweet = service.postTweet(text, latitude, longitude);
         } catch (IOException e) {
             e.printStackTrace();
         } catch (URISyntaxException e) {
             e.printStackTrace();
+        } catch (InputMismatchException e) {
+            System.err.println(TwitterUtil.INVALID_GEO_FORMAT);
         }
         if (tweet != null)
             System.out.printf(TwitterUtil.POST_SUCCESS_MSG_FORMAT, tweet.getIdStr());
         return tweet;
     }
 
-    private Scanner getParser(String[] args) {
+    private Scanner getCoordinateParser(String[] args) {
         if (coordinateParser == null)
-            coordinateParser = new Scanner(args[TwitterUtil.CMD_ARG_INDEX]).useDelimiter(COLON);
+            coordinateParser = new Scanner(args[TwitterUtil.CMD_DEFAULT_ARG_INDEX +1]);
+        coordinateParser.useDelimiter(COLON);
         return coordinateParser;
     }
 }
